Fixed stuff, that was requested while testing application
Admin can now view progress of individual users.
Added progress spinners when calculating correction progress.
Courses and submissions get now filterd by visibility for each user.
Replaced toggel-button in user-profile with tabs.
When creating new sheet, submission-time gets now automatically set to 23:55.
Added confirmation-dialogs when deleting stuff.
Added slider and shift-key-selection when assigning submissions to user.
Adjusted user-model in server to not require password, as it prohibited updating a user without changing the password (still required in front-end when adding user).

+++ client/src/app/courses/courses.component.ts
@@ -52,7 +52,9 @@ export class CoursesComponent implements OnInit {
   }
 
   delete(course: Course): void {
-    this.courses = this.courses.filter(c => c !== course);
-    this.courseService.deleteCourse(course).subscribe();
+    if (window.confirm('Wollen Sie den Kurs wirklich löschen?')) {
+      this.courses = this.courses.filter(c => c !== course);
+      this.courseService.deleteCourse(course).subscribe();
+    }
   }
 }
